pyproject.toml
[tool.poetry]
name = "gps"
version = "0.1.0"
description = ""
authors = ["user <panpanzov.nur@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
